chore: plan utils refactoring
