+++ frontend/src/components/Plan/logic/initialize/create.ts
@@ -7,7 +7,6 @@ import {
   useEnvironmentV1Store,
   useProjectV1Store,
   useSheetV1Store,
-  useStorageStore,
   batchGetOrFetchDatabases,
 } from "@/store";
 import { projectNamePrefix } from "@/store/modules/v1/common";
@@ -29,6 +28,7 @@ import {
 } from "@/utils";
 import { databaseEngineForSpec, sheetNameForSpec } from "../plan";
 import { createEmptyLocalSheet, getLocalSheetByName } from "../sheet";
+import { extractInitialSQLFromQuery } from "./util";
 
 export type InitialSQL = {
   sqlMap?: Record<string, string>;
@@ -319,32 +319,6 @@ const maybeSetInitialSQLForSpec = (
   }
 };
 
-const extractInitialSQLFromQuery = async (
-  query: Record<string, string>
-): Promise<InitialSQL> => {
-  const storageStore = useStorageStore();
-  const sql = query.sql;
-  if (sql && typeof sql === "string") {
-    return {
-      sql,
-    };
-  }
-  const sqlStorageKey = query.sqlStorageKey;
-  if (sqlStorageKey && typeof sqlStorageKey === "string") {
-    const sql = (await storageStore.get(sqlStorageKey)) || "";
-    return {
-      sql,
-    };
-  }
-  const sqlMapStorageKey = query.sqlMapStorageKey;
-  if (sqlMapStorageKey && typeof sqlMapStorageKey === "string") {
-    const sqlMap =
-      (await storageStore.get<Record<string, string>>(sqlMapStorageKey)) || {};
-    return { sqlMap };
-  }
-  return {};
-};
-
 const hasInitialSQL = (initialSQL?: InitialSQL) => {
   if (!initialSQL) {
     return false;
@@ -356,29 +330,4 @@ const hasInitialSQL = (initialSQL?: InitialSQL) => {
     return true;
   }
   return false;
-};
-
-export const isValidSpec = (spec: Plan_Spec): boolean => {
-  if (spec.changeDatabaseConfig) {
-    const sheetName = sheetNameForSpec(spec);
-    if (!sheetName) {
-      return false;
-    }
-    const uid = extractSheetUID(sheetName);
-    if (uid.startsWith("-")) {
-      const sheet = getLocalSheetByName(sheetName);
-      if (getSheetStatement(sheet).length === 0) {
-        return false;
-      }
-    } else {
-      const sheet = useSheetV1Store().getSheetByName(sheetName);
-      if (!sheet) {
-        return false;
-      }
-      if (getSheetStatement(sheet).length === 0) {
-        return false;
-      }
-    }
-  }
-  return true;
 };
